requirements.txt
python-dotenv==1.0.1
requests==2.32.3
peewee==3.17.5
pyTelegramBotAPI==4.19.1
setuptools==65.5.1
python-telegram-bot-pagination==0.0.3